Create birthday_paradox
I adapted a couple of things from the original code. 1) I used f-strings throughout and 2) used a leap year instead of a non-leap year to make sure all possible days were represented. This was good for me to do this instead of copying. The for loop to display the birthdays doesn't quite work for some reason and was solved just formatting the output more clearly.